requirements.txt
flask==2.0.1
pillow==9.0.0
numpy==1.21.0
openai>=1.0.0 